add attachemnt's new fields

+++ model/attachment.py
@@ -2,6 +2,7 @@ from flask import Blueprint, request, send_file
 
 from mongo import *
 from mongo import engine
+from mongo import User
 from .auth import *
 from .course import *
 from .utils import *
@@ -10,9 +11,9 @@ __all__ = ['attachment_api']
 attachment_api = Blueprint('attachment_api', __name__)
 
 
-@attachment_api.route('/<filename>', methods=['GET'])
+@attachment_api.route('/<id>', methods=['GET'])
 @login_required
-@Request.doc('filename', 'attachment', Attachment)
+@Request.doc('id', 'attachment', Attachment)
 def get_attachment(user, attachment):
     return send_file(
         attachment.file,
@@ -29,7 +30,11 @@ def get_attachment_list(user):
         'get all attachments\' names',
         data=[{
             'filename': a.filename,
-            'description': a.description
+            'description': a.description,
+            'author': a.author,
+            'created': a.created,
+            'updated': a.updated,
+            'id': a.id
         } for a in engine.Attachment.objects],
     )
 
@@ -49,7 +54,7 @@ def add_attachment(
     add an attachment to db
     '''
     try:
-        Attachment.add(file_obj, filename=filename, description=description)
+        Attachment.add(user, file_obj, filename=filename, description=description)
     except FileExistsError as e:
         return HTTPError(e, 400)
     except FileNotFoundError as e:
@@ -59,11 +64,11 @@ def add_attachment(
     return HTTPResponse('success')
 
 
-@attachment_api.route('/<filename>', methods=['PUT'])
+@attachment_api.route('/<id>', methods=['PUT'])
 @Request.files('file_obj')
 @Request.form('description')
 @identity_verify(0, 1)
-@Request.doc('filename', 'atta', Attachment)
+@Request.doc('id', 'atta', Attachment)
 def edit_attachment(
     user,
     file_obj,
@@ -73,6 +78,9 @@ def edit_attachment(
     '''
     update an attachment
     '''
+    user = User(user)
+    if atta.author != user and user < 'admin':
+        return HTTPError('Permission denied.', 403)
     try:
         atta.update(file_obj, description)
     except ValidationError as ve:
@@ -80,9 +88,9 @@ def edit_attachment(
     return HTTPResponse('success')
 
 
-@attachment_api.route('/<filename>', methods=['DELETE'])
+@attachment_api.route('/<id>', methods=['DELETE'])
 @identity_verify(0, 1)
-@Request.doc('filename', 'atta', Attachment)
+@Request.doc('id', 'atta', Attachment)
 def delete_attachment(
     user,
     atta,
@@ -90,5 +98,8 @@ def delete_attachment(
     '''
     delete an attachment
     '''
+    user = User(user)
+    if atta.author != user and user < 'admin':
+        return HTTPError('Permission denied.', 403)
     atta.delete()
     return HTTPResponse('success')
